refactor: 역직렬화 클래스를 response 내부 클래스로 변경
