+++ src/main/java/hiyen/galmanhae/dataprocess/application/CongestionService.java
@@ -1,7 +1,7 @@
 package hiyen.galmanhae.dataprocess.application;
 
 import hiyen.galmanhae.dataprocess.client.CongestionClient;
-import hiyen.galmanhae.dataprocess.dto.CongestionResponse;
+import hiyen.galmanhae.dataprocess.client.response.CongestionResponse;
 import hiyen.galmanhae.place.domain.vo.Congestion;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
