--- src/main/java/hiyen/galmanhae/dataprocess/dto/WeatherResponse.java
@@ -1,13 +1,0 @@
-package hiyen.galmanhae.dataprocess.dto;
-
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import hiyen.galmanhae.dataprocess.json.WeatherDeserializer;
-
-@JsonDeserialize(using = WeatherDeserializer.class)
-public record WeatherResponse(
-
-	String temperature,
-	String rainingProbability
-) {
-
-}
